refactor(voting option): refine voting option component to extract reusable sub components
- separate voting option to card container component and content by each voting type
